requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Data processing
Pillow>=9.0.0
tqdm>=4.65.0

# Scientific computing
matplotlib>=3.7.0
pandas>=2.0.0

# Performance monitoring
psutil>=5.9.0

# Acceleration (optional but recommended)
numba>=0.57.0

# Development tools
jupyter>=1.0.0
